requirements.txt
appdirs==1.4.4
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cssselect==1.1.0
dataclasses==0.8
fake-useragent==0.1.11
filelock==3.4.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
huggingface-hub==0.1.2
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.6.4
MarkupSafe==2.0.1
numpy==1.19.5
packaging==21.3
parse==1.19.0
pyee==8.2.2
PyJWT==2.3.0
pyparsing==3.0.6
pyppeteer==0.2.6
pyquery==1.4.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-html==0.10.0
sacremoses==0.0.46
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.27
tokenizers==0.10.3
torch==1.10.0
tqdm==4.62.3
transformers==4.12.5
typing_extensions==4.0.0
urllib3==1.26.7
w3lib==1.22.0
websockets==9.1
Werkzeug==2.0.2
zipp==3.6.0
